content-shield/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "content-shield"
version = "0.1.0"
description = "AI content quality validation framework with brand voice enforcement, factual accuracy checking, and resilience patterns."
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Content Shield Contributors"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.25",
    "tenacity>=8.0",
    "structlog>=23.0",
]

[project.optional-dependencies]
gcp = ["google-cloud-pubsub>=2.0", "google-cloud-bigquery>=3.0", "google-cloud-functions-framework>=3.0"]
slack = ["slack-sdk>=3.0"]
agents = ["google-generativeai>=0.3", "anthropic>=0.20", "openai>=1.0"]
dashboard = ["grafanalib>=0.7"]
all = ["content-shield[gcp,slack,agents,dashboard]"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.0",
    "pre-commit>=3.0",
]

[project.urls]
Homepage = "https://github.com/content-shield/content-shield"
Documentation = "https://content-shield.readthedocs.io"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
